Moved file operations to shared instead of repository (it does not store or read)
