+++ tests/e2e/mode_clean_all_test.py
@@ -5,7 +5,7 @@ import pytest
 
 from xmipp3_installer.application.cli.arguments import modes
 from xmipp3_installer.installer import constants
-from xmipp3_installer.repository import file_operations
+from xmipp3_installer.shared import file_operations
 
 from .. import get_assertion_message
