+++ tests/e2e/mode_config_test.py
@@ -6,8 +6,8 @@ import pytest
 
 from xmipp3_installer.application.cli.arguments import modes
 from xmipp3_installer.installer import constants
-from xmipp3_installer.repository import file_operations
 from xmipp3_installer.repository.config import ConfigurationFileHandler
+from xmipp3_installer.shared import file_operations
 
 from .. import (
 	get_file_content, normalize_file_line_endings,
